proto-loader: rename grpc-protobufjs to match npm
This commit renames grpc-protobufjs to proto-loader to more
closely match the name used on npm.
