+++ README.md
@@ -27,7 +27,7 @@ This library implements the core functionality of gRPC purely in JavaScript, wit
 
 ### gRPC Protobuf Loader
 
-Directory: [`packages/grpc-protobufjs`](https://github.com/grpc/grpc-node/tree/master/packages/grpc-protobufjs)
+Directory: [`packages/proto-loader`](https://github.com/grpc/grpc-node/tree/master/packages/proto-loader)
 
 npm package: [@grpc/proto-loader](https://www.npmjs.com/package/@grpc/proto-loader)
