+++ test/api/error_test.js
@@ -10,7 +10,7 @@ const _ = require('lodash');
 const anyGrpc = require('../any_grpc');
 const clientGrpc = anyGrpc.client;
 const serverGrpc = anyGrpc.server;
-const protoLoader = require('../../packages/grpc-protobufjs', options);
+const protoLoader = require('../../packages/proto-loader', options);
 const testServiceDef = protoLoader.loadSync(__dirname + '/../proto/test_service.proto');
 const TestServiceClient = clientGrpc.loadPackageDefinition(testServiceDef).TestService;
 const clientInsecureCreds = clientGrpc.credentials.createInsecure();
