+++ test/client-libraries-integration/use-grpc-js.js
@@ -4,7 +4,7 @@ const shimmer = require('shimmer');
 
 const grpcPJson = require('../../packages/grpc-js/package');
 const grpcImpl = require('../../packages/grpc-js');
-const grpcProtobuf = require('../../packages/grpc-protobufjs');
+const grpcProtobuf = require('../../packages/proto-loader');
 
 if (!process.env.USE_GRPC_NATIVE) {
   shimmer.wrap(Module, '_load', (moduleLoad) => {
